+++ src/app/@modal/(.)discover/loading.tsx
@@ -1,0 +1,3 @@
+'use client';
+
+export { default } from '@/app/(main)/discover/(detail)/loading';